src/main/java/com/monee/model/Reply.java
package com.monee.model;

import java.time.LocalDateTime;
import java.util.Objects;

public class Reply {

    private Long seq;
    private String content;
    private Account author;
    private LocalDateTime createAt;
    private LocalDateTime updateAt;
    private Post post;

    public Post getPost() {
        return post;
    }

    public void setPost(Post post) {
        this.post = post;
    }

    public Reply(String content){
        this.content = content;
    }

    public Long getSeq() {
        return seq;
    }

    public void setSeq(Long seq) {
        this.seq = seq;
    }


    public String getContent() {
        return content;
    }

    public void setContent(String content) {
        this.content = content;
    }

    public Account getAuthor() {
        return author;
    }

    public void setAuthor(Account author) {
        this.author = author;
    }

    public LocalDateTime getCreateAt() {
        return createAt;
    }

    public void setCreateAt(LocalDateTime createAt) {
        this.createAt = createAt;
    }

    public LocalDateTime getUpdateAt() {
        return updateAt;
    }

    public void setUpdateAt(LocalDateTime updateAt) {
        this.updateAt = updateAt;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Reply reply = (Reply) o;
        return seq.equals(reply.seq);
    }

    @Override
    public int hashCode() {
        return Objects.hash(seq);
    }

    @Override
    public String toString() {
        return "Reply{" +
                "seq=" + seq +
                ", content='" + content + '\'' +
                ", author=" + author +
                ", createAt=" + createAt +
                ", updateAt=" + updateAt +
                ", post=" + post +
                '}';
    }
}
